// WADNR.Web/src/app/shared/generated/model/gis-bulk-import-request.ts
export class GisBulkImportRequest { 
    ProjectIdentifierMetadataAttributeID?: number;
    ProjectNameMetadataAttributeID?: number;
    TreatmentTypeMetadataAttributeID?: number | null;
    CompletionDateMetadataAttributeID?: number | null;
    StartDateMetadataAttributeID?: number | null;
    ProjectStageMetadataAttributeID?: number | null;
    LeadImplementerMetadataAttributeID?: number | null;
    FootprintAcresMetadataAttributeID?: number | null;
    PrivateLandownerMetadataAttributeID?: number | null;
    TreatmentDetailedActivityTypeMetadataAttributeID?: number | null;
    TreatedAcresMetadataAttributeID?: number | null;
    PruningAcresMetadataAttributeID?: number | null;
    ThinningAcresMetadataAttributeID?: number | null;
    ChippingAcresMetadataAttributeID?: number | null;
    MasticationAcresMetadataAttributeID?: number | null;
    GrazingAcresMetadataAttributeID?: number | null;
    LopScatAcresMetadataAttributeID?: number | null;
    BiomassRemovalAcresMetadataAttributeID?: number | null;
    HandPileAcresMetadataAttributeID?: number | null;
    HandPileBurnAcresMetadataAttributeID?: number | null;
    MachinePileBurnAcresMetadataAttributeID?: number | null;
    BroadcastBurnAcresMetadataAttributeID?: number | null;
    OtherAcresMetadataAttributeID?: number | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}