// WADNR.Web/src/app/shared/generated/model/gis-feature-grid-row.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class GisFeatureGridRow { 
    GisFeatureID?: number;
    GisImportFeatureKey?: number;
    IsValid?: boolean | null;
    CalculatedArea?: number | null;
    MetadataValues?: { [key: string]: string | null; } | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface GisFeatureGridRowForm { 
    GisFeatureID?: FormControl<number>;
    GisImportFeatureKey?: FormControl<number>;
    IsValid?: FormControl<boolean>;
    CalculatedArea?: FormControl<number>;
    MetadataValues?: FormControl<{ [key: string]: string | null; }>;
}

export class GisFeatureGridRowFormControls { 
    public static GisFeatureID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static GisImportFeatureKey = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static IsValid = (value: FormControlState<boolean> | boolean = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<boolean>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static CalculatedArea = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static MetadataValues = (value: FormControlState<{ [key: string]: string | null; }> | { [key: string]: string | null; } = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<{ [key: string]: string | null; }>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}